+++ NEWS.md
@@ -27,6 +27,7 @@
 * `writeData()` calls `force(x)` to evaluate the object before options are set ([#264](https://github.com/ycphs/openxlsx/issues/264))
 * `createComment()` now correctly handles `integers` in `width` and `height` ([#275](https://github.com/ycphs/openxlsx/issues/275))
 * `setStyles()` accepts `halign="justify"` ([#305](https://github.com/ycphs/openxlsx/issues/305))
+* `data.frame`s with `NA` column names are converted to `"NA"` when saving ([#292](https://github.com/ycphs/openxlsx/issues/292), [#316](https://github.com/ycphs/openxlsx/issues/316)
 
 # openxlsx 4.2.4